social_service/requirements.txt
requests
Django==5.2.3
redis
psycopg2-binary
djangorestframework
djoser
djangorestframework-simplejwt